+++ NEWS.md
@@ -2,9 +2,16 @@
 
 # alphavantagepf 0.8.2: in Progress
 
-Bug todo:
+CUrrent: 0.8.136:  Adding tickerlists.
+
 
 New feature todo
+
+* Redo: 
+* Search tickerlist using internal data
+* Add dividends and earnings events to price graph
+* Replace symbol_search with search of listings data.table.  Will make MUCH faster
+
 
 * Total return and basic statistics with graphed prices.
 * Saving dividends and Earnings into database.
@@ -16,6 +23,7 @@ New feature todo
 
 New Features
 
+* `av_get_pf()` returns a date column for melted data.
 * `av_runShiny()` now returns a `ShinyAppHandle` object and does not block execution
 * `av_extract_df` now returns an empty data.table if there is nothing to extract. Set `empty_dt_onerror=FALSE` to throw an error.
 * Scatter plot added to ActiveTS, with piecewise linear regression, accessed via "tailhedge" scatter plot option
@@ -32,6 +40,8 @@ Small Changes
 * Improved filtering and search acrosss tables.
 * Option sets and parameters are now persistent
 * Copy Table just copies results from actions taken on Line 1
+
+Bug todo: ------------------------------------------
 
 Bug fixes